third_party/WebKit/LayoutTests/svg/custom/getBBox-use.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<svg height="0">
  <rect id="r" width="300" height="400"/>
  <use xlink:href="#r" x="100" y="200"/>
</svg>
<script>
test(function() {
    var useBBox = document.querySelector('use').getBBox();
    assert_equals(useBBox.x, 100);
    assert_equals(useBBox.y, 200);
    assert_equals(useBBox.width, 300);
    assert_equals(useBBox.height, 400);
}, 'Additional translation affects bounding box of <use>');
</script>
